% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_region.R
\name{add_region}
\alias{add_region}
\title{Add regions to pubDashboard dataframe}
\usage{
add_region(data, progress_bar = FALSE)
}
\arguments{
\item{data}{The dataframe on which to add region.}

\item{progress_bar}{Logical, whether to print a progress bar.}
}
\description{
Add regions to pubDashboard dataframe
}
\examples{
\dontrun{
x <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1, per_page = 1)
x <- add_region(x)
names(x)
}
}
