% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dygraph_year.R
\name{dygraph_year}
\alias{dygraph_year}
\title{Generate a dygraph of journal paper percentages, by country and year}
\usage{
dygraph_year(data, level = "continent")
}
\arguments{
\item{data}{The processed dataframe of data}

\item{level}{Level of analysis, either country or continent}
}
\description{
Generate a dygraph of journal paper percentages, by country and year
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
dygraph_year(data)
dygraph_year(data, "country")
}
}
