% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_openalex_pubs.R
\name{fetch_openalex_pubs}
\alias{fetch_openalex_pubs}
\title{Downloads relevant publication data using \code{openalexR}}
\usage{
fetch_openalex_pubs(journal = NULL, ...)
}
\arguments{
\item{journal}{The list of desired journals.}

\item{...}{Arguments passed to \code{\link[openalexR:oa_fetch]{openalexR::oa_fetch()}}}
}
\description{
Downloads relevant publication data using \code{openalexR}
}
\details{
As recommended by the authors of the \code{openalexR} package,

\emph{Before we go any further, we highly recommend you set \code{openalexR.mailto}
option so that your requests go to the polite pool for faster response times.
If you have OpenAlex Premium, you can add your API key to the
\code{openalexR.apikey} option as well. These lines best go into .Rprofile with
\code{file.edit("~/.Rprofile")}.}

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(openalexR.mailto = "example@email.com")
options(openalexR.apikey = "EXAMPLE_APIKEY")"
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{

x <- fetch_openalex_pubs(journal = "Collabra", pages = 1, per_page = 1)
names(x)
}
}
