% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_journal_id.R
\name{get_journal_id}
\alias{get_journal_id}
\title{Extract \code{openalexR} journal ID from journal names}
\usage{
get_journal_id(journal_name, verbose = TRUE)
}
\arguments{
\item{journal_name}{The list of desired journals (by journal name).}

\item{verbose}{Passed to \code{\link[openalexR:oa_fetch]{openalexR::oa_fetch()}} and defaults to \code{TRUE.}}
}
\description{
Extract \code{openalexR} journal ID from journal names
}
\details{
As recommended by the authors of the \code{openalexR} package,

\emph{Before we go any further, we highly recommend you set \code{openalexR.mailto}
option so that your requests go to the polite pool for faster response times.
If you have OpenAlex Premium, you can add your API key to the
\code{openalexR.apikey} option as well. These lines best go into .Rprofile with
\code{file.edit("~/.Rprofile")}.}

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(openalexR.mailto = "example@email.com")
options(openalexR.apikey = "EXAMPLE_APIKEY")"
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
x <- get_journal_id(journal_name = "Collabra")
x
x <- get_journal_id(journal_name = c(
  "Social Psychological and Personality Science", "Nature Human Behaviour"))
x$id
}
}
