% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapply_fetch_openalex_pubs.R
\name{lapply_fetch_openalex_pubs}
\alias{lapply_fetch_openalex_pubs}
\title{Loop \code{fetch_openalex_pubs} over journals or years}
\usage{
lapply_fetch_openalex_pubs(
  years = 1987:2023,
  journal_id,
  over = "year",
  from_publication_date = "1987-01-01",
  save = TRUE,
  file_suffix = "",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{years}{Desired list to loop over (for option \code{over = "year"}).}

\item{journal_id}{The list of desired journals (by OpenAlex ID).}

\item{over}{Looping over what. Options are \code{"year"} or \code{"journal"}.}

\item{from_publication_date}{Start date (for option \code{over = "journal"}).}

\item{save}{Whether to save the data to disk.}

\item{file_suffix}{What suffix to add to the file (e.g., "_new").}

\item{verbose}{Passed to \code{\link[openalexR:oa_fetch]{openalexR::oa_fetch()}} and defaults to \code{TRUE.}}

\item{...}{Arguments passed to \code{\link[openalexR:oa_fetch]{openalexR::oa_fetch()}}}
}
\description{
Loop \code{fetch_openalex_pubs} over journals or years
}
\details{
As recommended by the authors of the \code{openalexR} package,

\emph{Before we go any further, we highly recommend you set \code{openalexR.mailto}
option so that your requests go to the polite pool for faster response times.
If you have OpenAlex Premium, you can add your API key to the
\code{openalexR.apikey} option as well. These lines best go into .Rprofile with
\code{file.edit("~/.Rprofile")}.}

\if{html}{\out{<div class="sourceCode">}}\preformatted{options(openalexR.mailto = "example@email.com")
options(openalexR.apikey = "EXAMPLE_APIKEY")"
}\if{html}{\out{</div>}}
}
\examples{
\dontrun{
lapply_fetch_openalex_pubs(
  over = "journal",
  journal_id = c("https://openalex.org/S90392387",
                 "https://openalex.org/S33443600"),
 from_publication_date = "2024-01-01",
 to_publication_date = "2024-10-15",
 save = FALSE)
}
}
