% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dashboard.R
\name{render_dashboard}
\alias{render_dashboard}
\title{Render complete pubDashboard dashboard}
\usage{
render_dashboard(
  file_name = "dashboard",
  title = "title",
  author = "author",
  journal_name = NULL,
  journal_id = NULL,
  data_folder = "data",
  tab_continent = TRUE,
  tab_continent_year = TRUE,
  tab_continent_journal = TRUE,
  tab_country = TRUE,
  tab_country_journal = TRUE,
  tab_psychology = FALSE,
  tab_economics = FALSE,
  tab_general = FALSE,
  tab_figure1 = FALSE,
  tab_missing = TRUE,
  ...
)
}
\arguments{
\item{file_name}{Desired file name.}

\item{title}{Desired dashboard title.}

\item{author}{Desired displayed dashboard author.}

\item{journal_name}{The list of desired journals (by journal name).}

\item{journal_id}{The list of desired journals (by OpenAlex ID).}

\item{data_folder}{Folder where to save the data.}

\item{tab_continent}{Whether to render the "Continent" tab.}

\item{tab_continent_year}{Whether to render the "Continent by year" tab.}

\item{tab_continent_journal}{Whether to render the "Continent by journal" tab.}

\item{tab_country}{Whether to render the "Country" tab.}

\item{tab_country_journal}{Whether to render the "Country by journal" tab.}

\item{tab_psychology}{Whether to render the "Psychology" tab.}

\item{tab_economics}{Whether to render the "Economics" tab.}

\item{tab_general}{Whether to render the "General" tab.}

\item{tab_figure1}{Whether to render the "Figure 1" tab.}

\item{tab_missing}{Whether to render the "Missing" tab.}

\item{...}{Arguments passed to \code{\link[openalexR:oa_fetch]{openalexR::oa_fetch()}}}
}
\description{
Render complete pubDashboard dashboard
}
\examples{
\dontshow{
.old_wd <- setwd(tempdir())
}
\dontrun{
render_dashboard(
  file_name = "my_dashboard",
  title = "Wonderful Dashboard",
  author = "Rémi Thériault",
  journal_name = c("Journal of Personality and Social Psychology", "Health Psychology"),
  from_publication_date = "2024-01-01",
  tab_figure1 = TRUE
)
}
\dontshow{
unlink("data/data_new.rds")
setwd(.old_wd)
}
}
