% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_country_year.R
\name{scatter_country_year}
\alias{scatter_country_year}
\title{Generate table of journal paper percentages, by continent and year}
\usage{
scatter_country_year(
  data,
  method = "lm",
  ymin = 0,
  ymax = 100,
  yby = 20,
  plotly = TRUE,
  citation = NULL,
  citation_size = 15,
  text_size = NULL,
  height = NULL,
  ...
)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{method}{Which method to use for the regression line, either "lm" (default) or "loess".}

\item{ymin}{Minimum value for y-axis}

\item{ymax}{Maximum value for y-axis}

\item{yby}{Tick increments for y-axis}

\item{plotly}{Logical, whether to use plotly for dynamic data visualization.}

\item{citation}{Optionally, a citation to add as a footer.}

\item{citation_size}{Font size of the citation.}

\item{text_size}{Size of the element_text ggplot2 element}

\item{height}{Height argument of \link[plotly:ggplotly]{plotly::ggplotly}}

\item{...}{Further arguments passed to \link[rempsyc:nice_scatter]{rempsyc::nice_scatter}}
}
\description{
Generate table of journal paper percentages, by continent and year
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
suppressWarnings(scatter_country_year(data))
}
}
