% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_figure1.R
\name{scatter_figure1}
\alias{scatter_figure1}
\title{Generate table of journal paper percentages, by continent and year}
\usage{
scatter_figure1(data, method = "lm", original = TRUE, plotly = TRUE, ...)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{method}{Which method to use for the regression line, either "lm" (default) or "loess".}

\item{original}{Logical; if \code{TRUE}, attempts to mimic Arnett's (2008) Figure 1 in style.}

\item{plotly}{Logical, whether to use plotly for dynamic data visualization.}

\item{...}{Further arguments passed to \link[rempsyc:nice_scatter]{rempsyc::nice_scatter}}
}
\description{
Generate table of journal paper percentages, by continent and year
}
\examples{
\dontrun{
journals <- c("Developmental Psychology",
             "Journal of Personality and Social Psychology",
             "Journal of Abnormal Psychology",
             "Journal of Family Psychology",
             "Health Psychology",
             "Journal of Educational Psychology"
)

data <- fetch_openalex_pubs(journal_name = journals, pages = 10)
data <- clean_journals_continents(data)
scatter_figure1(data)
}
}
