% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_journal_count.R
\name{table_journal_count}
\alias{table_journal_count}
\title{Count number of papers per journal, with year range}
\usage{
table_journal_count(data, datatable = TRUE)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{datatable}{Whether to output a \link[DT:datatable]{DT::datatable} HTML table widget
instead of a regular dataframe (defaults to TRUE).}
}
\description{
Count number of papers per journal, with year range
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
table_journal_count(data)
}
}
