% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_missing_country.R
\name{table_missing_country}
\alias{table_missing_country}
\title{Generate table of journal paper percentages, by country}
\usage{
table_missing_country(data, datatable = TRUE)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{datatable}{Whether to output a \link[DT:datatable]{DT::datatable} HTML table widget
instead of a regular dataframe (defaults to TRUE).}
}
\description{
Generate table of journal paper percentages, by country
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
data[1, c(4, 6)] <- NA
table_missing_country(data)
}
}
