% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_continent.R
\name{waffle_continent}
\alias{waffle_continent}
\title{Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)}
\usage{
waffle_continent(
  data,
  citation = NULL,
  citation_size = NULL,
  data_formatted = NULL
)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{citation}{Optionally, a citation to add as a footer.}

\item{citation_size}{Font size of the citation.}

\item{data_formatted}{To save time, it is possible to reuse the dataframe
used for this function that is generated by \code{\link[=table_continent]{table_continent()}}.}
}
\description{
Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Journal of Economic Psychology", pages = 1)
data <- clean_journals_continents(data)
waffle_continent(data)
}
}
