% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_country.R
\name{waffle_country}
\alias{waffle_country}
\title{Generate a waffle plot made of country flags}
\usage{
waffle_country(data, citation = NULL, citation_size = NULL)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{citation}{Optionally, a citation to add as a footer.}

\item{citation_size}{Font size of the citation.}
}
\description{
Generate a waffle plot made of country flags
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
waffle_country(data)
}
}
