% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/waffle_country_journal.R
\name{waffle_country_journal}
\alias{waffle_country_journal}
\title{Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)}
\usage{
waffle_country_journal(
  data,
  citation = NULL,
  citation_size = NULL,
  journal_abbreviation = TRUE,
  data_formatted = NULL
)
}
\arguments{
\item{data}{The processed dataframe of data}

\item{citation}{Optionally, a citation to add as a footer.}

\item{citation_size}{Font size of the citation.}

\item{journal_abbreviation}{Logical, whether to use the journal abbreviation
to fit the entire plot, otherwise some journal names can be quite long and
accordingly be cropped.}

\item{data_formatted}{To save time, it is possible to reuse the dataframe
used for this function that is generated by \code{\link[=table_country_journal]{table_country_journal()}}.}
}
\description{
Generate a waffle chart of journal paper percentages, by continent (each square = 1\% of data)
}
\examples{
\dontrun{
data <- fetch_openalex_pubs(journal_name = "Collabra", pages = 1)
data <- clean_journals_continents(data)
waffle_country_journal(data)
}
}
